# artifacts/open_api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Chat APIs
  version: 0.215.37
  description: API for the Simple Chat, handling chats, conversations, documents, user settings, groups, feedback, prompts, models, safety, and admin functions.
servers:
  - url: /api
    description: API Root
...
paths:
...
  /group_prompts:
    get:
      tags:
        - Group Prompts
        - Groups
      summary: Get prompts for the active group (paginated)
      description: Retrieves a paginated list of prompts associated with the user's currently active group. Requires `enable_group_workspaces` setting to be true and user membership in the group.
      security:
        - bearerAuth: []
      parameters:
        # Add pagination/filtering params based on list_prompts implementation if needed
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: The page number.
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
          description: Items per page.
        - name: search
          in: query
          schema:
            type: string
          description: Optional search term for prompt name.
      responses:
        '200':
          description: Group prompts retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupPrompt'
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_count:
                    type: integer
        '400':
          description: Bad Request (No active group selected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_group_workspaces` disabled or user not in group - implicit check)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database query failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Group Prompts
        - Groups
      summary: Create a new prompt for the active group
      description: Creates a new shared prompt within the user's active group. Requires `enable_group_workspaces` setting to be true and user membership in the group (permission to create might depend on role).
      security:
        - bearerAuth: [] # Role check logic inside if needed based on group policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - content
              properties:
                name:
                  type: string
                  description: The name/title of the prompt.
                content:
                  type: string
                  description: The text content of the prompt.
      responses:
        '201': # Code returns 201
          description: Group prompt created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPrompt' # Returns the created prompt
        '400':
          description: Bad Request (Missing name/content or no active group)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_group_workspaces` disabled or insufficient permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database create failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /group_prompts/{prompt_id}:
    get:
      tags:
        - Group Prompts
        - Groups
      summary: Get a specific group prompt
      description: Retrieves the details of a single prompt belonging to the user's active group. Requires `enable_group_workspaces` setting to be true and user membership in the group.
      security:
        - bearerAuth: []
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the group prompt to retrieve.
      responses:
        '200':
          description: Group prompt retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPrompt'
        '400':
          description: Bad Request (No active group selected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_group_workspaces` disabled or user not in group)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group prompt not found in the active group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database read failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Group Prompts
        - Groups
      summary: Update a group prompt
      description: Updates the name and/or content of a specific prompt within the user's active group. Requires `enable_group_workspaces` setting to be true and user membership (permission to update might depend on role or creator).
      security:
        - bearerAuth: [] # Permission check logic inside
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the group prompt to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1 # Must provide at least name or content
              properties:
                name:
                  type: string
                  description: The new name for the prompt.
                content:
                  type: string
                  description: The new content for the prompt.
      responses:
        '200':
          description: Group prompt updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPrompt' # Returns the updated prompt
        '400':
          description: Bad Request (No active group selected, invalid data, or no fields provided)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_group_workspaces` disabled or insufficient permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group prompt not found in the active group or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database update failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Group Prompts
        - Groups
      summary: Delete a group prompt
      description: Deletes a specific prompt from the user's active group. Requires `enable_group_workspaces` setting to be true and user membership (permission to delete might depend on role or creator).
      security:
        - bearerAuth: [] # Permission check logic inside
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the group prompt to delete.
      responses:
        '200':
          description: Group prompt deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Prompt deleted successfully"
        '400':
          description: Bad Request (No active group selected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Feature `enable_group_workspaces` disabled or insufficient permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group prompt not found in the active group or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Database delete failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --- Group Management Routes (route_backend_groups.py) ---
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error that occurred.
...
    GroupPrompt:
      type: object
      description: A prompt shared within a group.
      properties:
        id:
          type: string
          format: uuid
        group_id:
          type: string
          format: uuid
          description: ID of the group this prompt belongs to.
        user_id: # ID of the user who created/uploaded the prompt
          type: string
          description: ID of the user who created the prompt.
        name:
          type: string
          description: Name or title of the prompt.
        content:
          type: string
          description: The text content of the prompt.
        type:
          type: string
          example: "group_prompt"
          description: Internal type identifier.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the prompt was created.
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the prompt was last updated.
...
  securitySchemes:
    bearerAuth: # Defines how API clients authenticate
      type: http
      scheme: bearer
      bearerFormat: JWT # Indicates the type of token expected
      description: "Azure AD issued JWT Bearer token required for authenticated endpoints. Include in the request header as 'Authorization: Bearer <YOUR_TOKEN>'."